requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.3
django-cors==0.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
joblib==1.4.2
jwt==1.3.1
numpy==2.1.3
pycparser==2.22
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.5.2
threadpoolctl==3.5.0
tzdata==2024.2
